Dedicated composables for shootResources and shootExtensions
